Add StringExtensions.cs with various string manipulation methods
